Update README.md
# fedya

ABOUT TOOL :

This is a simple Python script that embeds images into an MP3 file.

This script will create a graphical user interface with two text fields and buttons for selecting an MP3 file and an image. Once the user chooses the desired files and clicks the "Embed Image" button, the script will process them and embed the image into the MP3 file. The result of the operation will be displayed in a label below the "Embed Image" button.

Before proceeding with the script, make sure you have the mutagen library installed. If not, install it using the following command:

pip install mutagen